+++ Tests/Unit/MSFT_PfxImport.Tests.ps1
@@ -51,19 +51,23 @@ try
         $validCertFullPath = '{0}\{1}' -f $validCertPath, $validThumbprint
 
         $validCertificateWithPrivateKey = @{
+            Thumbprint = $validThumbprint
             HasPrivateKey = $true
         }
 
         $validCertificateWithoutPrivateKey = @{
+            Thumbprint = $validThumbprint
             HasPrivateKey = $false
         }
 
-        $testCertificatePath_parameterfilter = {
+        $testPath_parameterfilter = {
             $Path -eq $validPath
         }
 
-        $testGetChildItem_parameterfilter = {
-            $Path -eq $validCertFullPath
+        $getCertificateFromCertificateStore_parameterfilter = {
+            $Thumbprint -eq $validThumbprint -and `
+            $Location -eq 'LocalMachine' -and `
+            $Store -eq 'My'
         }
 
         $importPfxCertificate_parameterfilter = {
@@ -71,6 +75,12 @@ try
                 $FilePath -eq $validPath -and `
                 $Exportable -eq $True -and `
                 $Password -eq $testCredential.Password
+        }
+
+        $removeCertificateFromCertificateStore_parameterfilter = {
+            $Location -eq 'LocalMachine' -and `
+            $Store -eq 'My' -and `
+            $Thumbprint -eq $validThumbprint
         }
 
         $presentParams = @{
@@ -93,20 +103,12 @@ try
         }
 
         Describe 'MSFT_PfxImport\Get-TargetResource' -Tag 'Get' {
-            Context 'When the PFX file exists and the certificate exists with a private key' {
+            Context 'When the certificate exists with a private key' {
                 Mock `
-                    -CommandName Test-CertificatePath `
+                    -CommandName Get-CertificateFromCertificateStore `
                     -MockWith {
-                    $true
-                } `
-                    -ParameterFilter $testCertificatePath_parameterfilter
-
-                Mock `
-                    -CommandName Get-ChildItem `
-                    -MockWith {
-                    $validCertificateWithPrivateKey
-                } `
-                    -ParameterFilter $testGetChildItem_parameterfilter
+                        $validCertificateWithPrivateKey
+                    }
 
                 $result = Get-TargetResource @presentParams
 
@@ -122,31 +124,19 @@ try
 
                 It 'Should call the expected mocks' {
                     Assert-MockCalled `
-                        -CommandName Test-CertificatePath `
-                        -ParameterFilter $testCertificatePath_parameterfilter `
-                        -Exactly -Times 1
-
-                    Assert-MockCalled `
-                        -CommandName Get-ChildItem `
-                        -ParameterFilter $testGetChildItem_parameterfilter `
+                        -CommandName Get-CertificateFromCertificateStore `
+                        -ParameterFilter $getCertificateFromCertificateStore_parameterfilter `
                         -Exactly -Times 1
                 }
             }
 
-            Context 'When the PFX file exists and the certificate exists without private key' {
+            Context 'When the certificate exists without private key' {
                 Mock `
-                    -CommandName Test-CertificatePath `
+                    -CommandName Get-CertificateFromCertificateStore `
                     -MockWith {
-                    $true
-                } `
-                    -ParameterFilter $testCertificatePath_parameterfilter
+                        $validCertificateWithoutPrivateKey
+                    }
 
-                Mock `
-                    -CommandName Get-ChildItem `
-                    -MockWith {
-                    $validCertificateWithoutPrivateKey
-                } `
-                    -ParameterFilter $testGetChildItem_parameterfilter
 
                 $result = Get-TargetResource @presentParams
 
@@ -162,28 +152,15 @@ try
 
                 It 'Should call the expected mocks' {
                     Assert-MockCalled `
-                        -CommandName Test-CertificatePath `
-                        -ParameterFilter $testCertificatePath_parameterfilter `
-                        -Exactly -Times 1
-
-                    Assert-MockCalled `
-                        -CommandName Get-ChildItem `
-                        -ParameterFilter $testGetChildItem_parameterfilter `
+                        -CommandName Get-CertificateFromCertificateStore `
+                        -ParameterFilter $getCertificateFromCertificateStore_parameterfilter `
                         -Exactly -Times 1
                 }
             }
 
-            Context 'When the PFX file exists and the certificate does not exist' {
+            Context 'When the certificate does not exist' {
                 Mock `
-                    -CommandName Test-CertificatePath `
-                    -MockWith {
-                    $true
-                } `
-                    -ParameterFilter $testCertificatePath_parameterfilter
-
-                Mock `
-                    -CommandName Get-ChildItem `
-                    -ParameterFilter $testGetChildItem_parameterfilter
+                    -CommandName Get-CertificateFromCertificateStore
 
                 $result = Get-TargetResource @presentParams
 
@@ -199,78 +176,8 @@ try
 
                 It 'Should call the expected mocks' {
                     Assert-MockCalled `
-                        -CommandName Test-CertificatePath `
-                        -ParameterFilter $testCertificatePath_parameterfilter `
-                        -Exactly -Times 1
-
-                    Assert-MockCalled `
-                        -CommandName Get-ChildItem `
-                        -ParameterFilter $testGetChildItem_parameterfilter `
-                        -Exactly -Times 1
-                }
-            }
-
-            Context 'When the PFX file does not exist and ensure is absent' {
-                Mock `
-                    -CommandName Test-CertificatePath `
-                    -MockWith {
-                    $false
-                } `
-                    -ParameterFilter $testCertificatePath_parameterfilter
-
-                Mock `
-                    -CommandName Get-ChildItem `
-                    -MockWith {
-                    $validCertificateWithoutPrivateKey
-                } `
-                    -ParameterFilter $testGetChildItem_parameterfilter
-
-                $result = Get-TargetResource @absentParams
-
-                It 'Should return a hashtable' {
-                    $result | Should -BeOfType System.Collections.Hashtable
-                }
-
-                It 'Should contain the input values' {
-                    $result.Thumbprint | Should -BeExactly $validThumbprint
-                    $result.Ensure | Should -BeExactly 'Absent'
-                }
-
-                It 'Should call the expected mocks' {
-                    Assert-MockCalled `
-                        -CommandName Test-CertificatePath `
-                        -ParameterFilter $testCertificatePath_parameterfilter `
-                        -Exactly -Times 0
-
-                    Assert-MockCalled `
-                        -CommandName Get-ChildItem `
-                        -ParameterFilter $testGetChildItem_parameterfilter `
-                        -Exactly -Times 1
-                }
-            }
-
-            Context 'When the PFX file does not exist and ensure is present' {
-                Mock `
-                    -CommandName Test-CertificatePath `
-                    -MockWith {
-                    $false
-                } `
-                    -ParameterFilter $testCertificatePath_parameterfilter
-
-                It 'Should throw expected exception' {
-                    $errorRecord = Get-InvalidArgumentRecord `
-                        -Message ($LocalizedData.CertificatePfxFileNotFoundError -f $validPath) `
-                        -ArgumentName 'Path'
-
-                    {
-                        $script:result = Get-TargetResource @presentParams
-                    } | Should -Throw $errorRecord
-                }
-
-                It 'Should call the expected mocks' {
-                    Assert-MockCalled `
-                        -CommandName Test-CertificatePath `
-                        -ParameterFilter $testCertificatePath_parameterfilter `
+                        -CommandName Get-CertificateFromCertificateStore `
+                        -ParameterFilter $getCertificateFromCertificateStore_parameterfilter `
                         -Exactly -Times 1
                 }
             }
@@ -289,7 +196,7 @@ try
                 Test-TargetResource @presentParams | Should -BeOfType Boolean
             }
 
-            Context 'When valid path and thumbprint and certificate is not in store but should be' {
+            Context 'When certificate is not in store but should be' {
                 It 'Should return false' {
                     Mock -CommandName Get-TargetResource {
                         return @{
@@ -303,7 +210,7 @@ try
                 }
             }
 
-            Context 'When valid path and thumbprint and PFX is not in store and should not be' {
+            Context 'When certificate is not in store and should not be' {
                 It 'Should return true' {
                     Mock -CommandName Get-TargetResource {
                         return @{
@@ -317,7 +224,7 @@ try
                 }
             }
 
-            Context 'When valid path and thumbprint and PFX is in store and should be' {
+            Context 'When certificate is in store and should be' {
                 It 'Should return true' {
                     Mock -CommandName Get-TargetResource {
                         return @{
@@ -331,7 +238,7 @@ try
                 }
             }
 
-            Context 'When valid path and thumbprint and PFX is in store but should not be' {
+            Context 'When certificate is in store but should not be' {
                 It 'Should return false' {
                     Mock -CommandName Get-TargetResource {
                         return @{
@@ -347,50 +254,81 @@ try
         }
 
         Describe 'MSFT_PfxImport\Set-TargetResource' -Tag 'Set' {
-            Context 'When PFX file exists and thumbprint and Ensure is Present' {
-                Mock `
-                    -CommandName Import-PfxCertificate `
-                    -ParameterFilter $importPfxCertificate_parameterfilter
-                Mock -CommandName Get-ChildItem
-                Mock -CommandName Remove-Item
+            BeforeAll {
+                Mock -CommandName Test-Path -MockWith { $true }
+                Mock -CommandName Import-PfxCertificate
+                Mock -CommandName Remove-CertificateFromCertificateStore
+            }
 
+            Context 'When PFX file exists and certificate should be in the store' {
                 Set-TargetResource @presentParams
 
-                It 'Should call Import-PfxCertificate with the parameters supplied' {
+                It 'Should call Test-Path with the parameters supplied' {
+                    Assert-MockCalled `
+                        -CommandName Test-Path `
+                        -ParameterFilter $testPath_parameterfilter `
+                        -Exactly -Times 1
+                }
+
+                It 'Should call Import-Certificate with the parameters supplied' {
                     Assert-MockCalled `
                         -CommandName Import-PfxCertificate `
                         -ParameterFilter $importPfxCertificate_parameterfilter `
                         -Exactly -Times 1
                 }
 
-                It 'Should not call Get-ChildItem' {
-                    Assert-MockCalled -CommandName Get-ChildItem -Exactly -Times 0
-                }
-
-                It 'Should not call Remove-Item' {
-                    Assert-MockCalled -CommandName Remove-Item -Exactly -Times 0
+                It 'Should not call Remove-CertificateFromCertificateStore' {
+                    Assert-MockCalled -CommandName Remove-CertificateFromCertificateStore -Exactly -Times 0
                 }
             }
 
-            Context 'When certificate exists and Ensure is Absent' {
-                Mock -CommandName Import-PfxCertificate
-                Mock -CommandName Get-ChildItem -MockWith {
-                    @{ Thumbprint = $validThumbprint }
-                }
-                Mock -CommandName Remove-Item
-
+            Context 'When PFX file exists and certificate should not be in the store' {
                 Set-TargetResource @absentParams
+
+                It 'Should call Test-Path with the parameters supplied' {
+                    Assert-MockCalled `
+                        -CommandName Test-Path `
+                        -ParameterFilter $testPath_parameterfilter `
+                        -Exactly -Times 0
+                }
 
                 It 'Should not call Import-PfxCertificate' {
                     Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
                 }
 
-                It 'Should call Get-ChildItem' {
-                    Assert-MockCalled -CommandName Get-ChildItem -Exactly -Times 1
+                It 'Should not call Remove-CertificateFromCertificateStore' {
+                    Assert-MockCalled `
+                        -CommandName Remove-CertificateFromCertificateStore `
+                        -ParameterFilter $removeCertificateFromCertificateStore_parameterfilter `
+                        -Exactly -Times 1
+                }
+            }
+
+            Context 'When PFX file does not exists and certificate should be in the store' {
+                Mock -CommandName Test-Path -MockWith { $false }
+
+                It 'Should throw exception' {
+                    {
+                        Set-TargetResource @presentParams
+                    } | Should -Throw ($script:localizedData.CertificatePfxFileNotFoundError -f $validPath)
                 }
 
-                It 'Should call Remove-Item' {
-                    Assert-MockCalled -CommandName Remove-Item -Exactly -Times 1
+                It 'Should call Test-Path with the parameters supplied' {
+                    Assert-MockCalled `
+                        -CommandName Test-Path `
+                        -ParameterFilter $testPath_parameterfilter `
+                        -Exactly -Times 1
+                }
+
+                It 'Should not call Import-PfxCertificate' {
+                    Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
+                }
+
+                It 'Should not call Remove-CertificateFromCertificateStore' {
+                    Assert-MockCalled `
+                        -CommandName Remove-CertificateFromCertificateStore `
+                        -ParameterFilter $removeCertificateFromCertificateStore_parameterfilter `
+                        -Exactly -Times 0
                 }
             }
         }