+++ CHANGELOG.md
@@ -14,10 +14,17 @@
 - Refactor unit tests to update to latest standards.
 - CertificateImport:
   - Refactor to use common functions and share more code with `PfxImport`
-    resource .
+    resource.
+  - Resource will now only throw an exception if the PFX file does not exist
+    and it needs to be imported.
+  - Removed file existence check from `Path` parameter to enable the resource
+    to remove a certificate from the store without the need to have the
+    access to the certificate file.
 - CertificatePfx:
   - Refactor to use common functions and share more code with
     `CertificateImport` resource.
+  - Resource will now only throw an exception if the certificate file does
+    not exist and it needs to be imported.
 
 ## 4.6.0.0